Control the component, don't let it control you
The picker is a controlled component. This means we expect the on*Change handlers to be in charge of updating our `selectedValue` prop. That way they can also disallow/undo/mutate the selection of certain values. In other words, the embedder of this component should be the source of truth, not the native component.

(Copied from PickerAndroid)